+++ cmd/compute/load_balancer/nlb.go
@@ -1,4 +1,4 @@
-package cmd
+package load_balancer
 
 import (
 	exocmd "github.com/exoscale/cli/cmd"
