+++ cmd/compute/load_balancer/nlb_service_update.go
@@ -1,4 +1,4 @@
-package cmd
+package load_balancer
 
 import (
 	"errors"
